[CI] Add initial GitHub Actions definitions
This commit adds the ability for GitHub actions to run on the project.
This adds two simple CI scripts. They are:
* black.yml - used for code formatting checks. Will run in Pull Requests
aswell as when new commits are added to the main branch.
* postmerge.yml - used to run the various unittests for the project.
These are run when new commits are added to the main branch of the
project.

Functionality for GitHub Actions running Unittests in Pull Requests
will come in a later commit.