[ci] Add regular testing pipeline
This pipeline will run at 0900GMT on a Monday, running the test suite
to scan LLVM's repository and return if there are issues. Currently,
the output will always lead to a pass. The tests need to be matured in
the tool to cause a failure when an incorrect instruction is found.