[v0.2] Add Support for Pytest
In v0.1, the tool simply printed for the user if the instruction
was valid or not. However, in CI this was found to ensure
the pipeline would always pass. The tool has now adopted
Pytest, meaning it can properly output failures for the user,
and fail pipelines when this run through GitHub Actions.

CI Pipelines and relevant documentation have been updated
to reflect the change.