[v0.3] Expand supported instructions
Following this commit, the tool now supports extra instructions. The
new instructions are listed in the Version History section of the
README. To support these new instructions, the
SupportedEarlyclobberInstructions class has been expanded with new
functions to ensure that the process is correct for the different
instructions, as they are not the same as one another in format or
supported elements of the instructions.

Unittests have been updated to support the new functions within the
class.